pyproject.toml
[project]
name = "twittermanager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=0.5.0",
    "python-dotenv>=1.0.1",
    "schedule>=1.2.2",
    "tweepy>=4.15.0",
]
